chore(slack_app): remove the dead personal-GitHub gate surface
The authorship gate's activities, its "Continue as PostHog" button, and the
interactivity routing behind it are all unreachable now that the workflow no
longer calls them.

The slack-app-bot-prs flag goes with them. It was never created, so it
evaluated false everywhere and the team install has never contributed repos
to a mention — repo resolution reads the mentioner's personal install and
nothing else. Removing the flag keeps that behavior and drops the branch that
never ran.
