+++ posthog/temporal/ai/slack_app/__init__.py
@@ -26,12 +26,10 @@ from posthog.temporal.ai.slack_app.activities import (
     handle_posthog_code_slack_mention_command_activity,
     mark_slack_app_message_processing_activity,
     mark_slack_app_message_queued_activity,
-    post_posthog_code_authorship_timeout_activity,
     post_posthog_code_internal_error_activity,
     post_posthog_code_no_repos_activity,
     post_posthog_code_picker_timeout_activity,
     post_posthog_code_repo_picker_activity,
-    resolve_posthog_code_authorship_activity,
     resolve_posthog_code_slack_command_user_activity,
     resolve_posthog_code_slack_user_activity,
     run_posthog_slack_inbox_onboarding_activity,
@@ -65,9 +63,7 @@ SLACK_APP_ACTIVITIES = [
     post_posthog_code_no_repos_activity,
     post_posthog_code_repo_picker_activity,
     block_posthog_code_task_if_no_personal_github_activity,
-    resolve_posthog_code_authorship_activity,
     post_posthog_code_picker_timeout_activity,
-    post_posthog_code_authorship_timeout_activity,
     post_posthog_code_internal_error_activity,
     create_posthog_code_task_for_repo_activity,
     forward_posthog_code_followup_activity,
@@ -114,12 +110,10 @@ __all__ = [
     "handle_posthog_code_slack_mention_command_activity",
     "mark_slack_app_message_processing_activity",
     "mark_slack_app_message_queued_activity",
-    "post_posthog_code_authorship_timeout_activity",
     "post_posthog_code_internal_error_activity",
     "post_posthog_code_no_repos_activity",
     "post_posthog_code_picker_timeout_activity",
     "post_posthog_code_repo_picker_activity",
-    "resolve_posthog_code_authorship_activity",
     "resolve_posthog_code_slack_command_user_activity",
     "resolve_posthog_code_slack_user_activity",
     "run_posthog_slack_inbox_onboarding_activity",
