+++ posthog/temporal/ai/slack_app/activities/__init__.py
@@ -14,12 +14,10 @@ from posthog.temporal.ai.slack_app.activities.messaging import (
     block_posthog_code_task_if_no_personal_github_activity,
     mark_slack_app_message_processing_activity,
     mark_slack_app_message_queued_activity,
-    post_posthog_code_authorship_timeout_activity,
     post_posthog_code_internal_error_activity,
     post_posthog_code_no_repos_activity,
     post_posthog_code_picker_timeout_activity,
     post_posthog_code_repo_picker_activity,
-    resolve_posthog_code_authorship_activity,
 )
 from posthog.temporal.ai.slack_app.activities.onboarding import (
     run_posthog_slack_inbox_onboarding,
@@ -68,12 +66,10 @@ __all__ = [
     "handle_posthog_code_slack_mention_command_activity",
     "mark_slack_app_message_processing_activity",
     "mark_slack_app_message_queued_activity",
-    "post_posthog_code_authorship_timeout_activity",
     "post_posthog_code_internal_error_activity",
     "post_posthog_code_no_repos_activity",
     "post_posthog_code_picker_timeout_activity",
     "post_posthog_code_repo_picker_activity",
-    "resolve_posthog_code_authorship_activity",
     "resolve_posthog_code_slack_command_user_activity",
     "resolve_posthog_code_slack_user_activity",
     "run_posthog_slack_inbox_onboarding",
