+++ posthog/temporal/tests/ai/test_block_missing_user_github.py
@@ -2,8 +2,6 @@ from unittest.mock import MagicMock, patch
 
 from django.test import TestCase
 
-from parameterized import parameterized
-
 from posthog.models.integration import Integration
 from posthog.models.organization import Organization
 from posthog.models.team.team import Team
@@ -112,36 +110,6 @@ class TestBlockPostHogCodeTaskIfNoPersonalGitHub(TestCase):
 
         assert blocked is True
         mock_slack.client.chat_postMessage.assert_called_once()
-
-    @parameterized.expand(
-        [
-            ("flag_on_with_team_install_proceeds", True, True, False),
-            ("flag_off_with_team_install_blocks", False, True, True),
-            ("flag_on_without_team_install_blocks", True, False, True),
-        ]
-    )
-    @patch("products.slack_app.backend.feature_flags.posthoganalytics.feature_enabled")
-    @patch("posthog.temporal.ai.slack_app.activities.messaging.SlackIntegration")
-    def test_allow_bot_prs_gates_on_flag_and_team_install(
-        self, _name, flag_on, has_team_install, expect_blocked, mock_slack_cls, mock_flag
-    ):
-        mock_flag.return_value = flag_on
-        if has_team_install:
-            Integration.objects.create(
-                team=self.team,
-                kind="github",
-                config={"account": {"name": "posthog"}},
-                sensitive_config={"access_token": "ghp-test"},
-            )
-        mock_slack = MagicMock()
-        mock_slack_cls.return_value = mock_slack
-
-        blocked = block_posthog_code_task_if_no_personal_github_activity(
-            _make_inputs(self.integration.id), "C123", "1234.5678", self.user.id, True
-        )
-
-        assert blocked is expect_blocked
-        assert mock_slack.client.chat_postMessage.called is expect_blocked
 
     @patch("posthog.temporal.ai.slack_app.activities.messaging.SlackIntegration")
     def test_another_users_github_integration_does_not_count(self, mock_slack_cls):
