+++ products/slack_app/backend/feature_flags.py
@@ -38,7 +38,6 @@ SLACK_APP_OAUTH_FLAG = "slack-app-oauth"
 SLACK_APP_HOME_FLAG = "slack-app-home"
 SLACK_APP_AGENT_DESIGN_FLAG = "slack-app-agent-design"
 SLACK_APP_ASSISTANT_FLAG = "slack-app-assistant"
-SLACK_APP_BOT_PRS_FLAG = "slack-app-bot-prs"
 SLACK_APP_LIVING_ARTIFACTS_FLAG = "slack-app-living-artifacts"
 SLACK_APP_CANVAS_FILE_ARTIFACTS_FLAG = "slack-app-canvas-file-artifacts"
 SLACK_APP_MODEL_CLASSIFIER_FLAG = "slack-app-model-classifier"
@@ -237,28 +236,6 @@ def is_slack_app_untagged_thread_followups_enabled(integration: Integration, sla
             slack_team_id=slack_team_id,
             integration_id=integration.id,
         )
-        return False
-
-
-def is_slack_app_bot_prs_enabled(team: Team) -> bool:
-    """Gate for the workspace's team GitHub install contributing repos alongside the
-    mentioner's personal one. A GitHub-side capability, so this is the flag alone."""
-    organization_id = str(team.organization_id)
-    project_id = str(team.id)
-    try:
-        return bool(
-            posthoganalytics.feature_enabled(
-                SLACK_APP_BOT_PRS_FLAG,
-                str(team.uuid),
-                groups={"organization": organization_id, "project": project_id},
-                group_properties={"organization": {"id": organization_id}, "project": {"id": project_id}},
-                person_properties=_region_properties(),
-                only_evaluate_locally=False,
-                send_feature_flag_events=False,
-            )
-        )
-    except Exception:
-        logger.exception("slack_app_bot_prs_flag_check_failed", team_id=team.id)
         return False
