+++ products/slack_app/backend/tests/test_guess_repository.py
@@ -132,111 +132,6 @@ class TestGetFullRepoNames:
         assert result == ["posthog/alpha", "posthog/middle", "posthog/zebra"]
 
 
-@patch("products.slack_app.backend.api.GitHubIntegration")
-@patch("products.slack_app.backend.api.UserGitHubIntegration")
-class TestGetFullRepoNamesBotPRs:
-    @pytest.fixture(autouse=True)
-    def setup(self, db):
-        cache.clear()
-        self.organization = Organization.objects.create(name="Bot Org")
-        self.team = Team.objects.create(organization=self.organization, name="Bot Team")
-        self.user = User.objects.create(email="bot@example.com", distinct_id="user-bot")
-        self.slack_integration = Integration.objects.create(
-            team=self.team,
-            kind="slack",
-            integration_id="T_BOT",
-            sensitive_config={"access_token": "xoxb-bot"},
-        )
-
-    def _create_team_github_integration(self) -> Integration:
-        return Integration.objects.create(
-            team=self.team,
-            kind="github",
-            integration_id="gh-team-1",
-            config={"account": {"name": "posthog"}},
-            sensitive_config={"access_token": "gh-team-token"},
-        )
-
-    @patch("products.slack_app.backend.api.posthoganalytics.feature_enabled", return_value=False)
-    def test_flag_off_excludes_team_repos(self, _mock_feature_enabled, mock_user_github_class, mock_team_github_class):
-        _create_user_github_integration(self.user)
-        self._create_team_github_integration()
-
-        mock_user_github = MagicMock()
-        mock_user_github.list_all_cached_repositories.return_value = [
-            _repo_dict("posthog", "posthog", 1),
-            _repo_dict("posthog", "posthog-js", 2),
-        ]
-        mock_user_github_class.return_value = mock_user_github
-
-        result = _get_full_repo_names(self.slack_integration, user_id=self.user.id)
-
-        # The team install is present but, with the flag off, must never be consulted.
-        assert result == ["posthog/posthog", "posthog/posthog-js"]
-        mock_team_github_class.assert_not_called()
-
-    @patch("products.slack_app.backend.api.posthoganalytics.feature_enabled", return_value=True)
-    def test_flag_on_no_team_install_returns_only_user_repos(
-        self, _mock_feature_enabled, mock_user_github_class, mock_team_github_class
-    ):
-        _create_user_github_integration(self.user)
-
-        mock_user_github = MagicMock()
-        mock_user_github.list_all_cached_repositories.return_value = [
-            _repo_dict("posthog", "posthog", 1),
-            _repo_dict("posthog", "posthog-js", 2),
-        ]
-        mock_user_github_class.return_value = mock_user_github
-
-        result = _get_full_repo_names(self.slack_integration, user_id=self.user.id)
-
-        assert result == ["posthog/posthog", "posthog/posthog-js"]
-        mock_team_github_class.assert_not_called()
-
-    @patch("products.slack_app.backend.api.posthoganalytics.feature_enabled", return_value=True)
-    def test_flag_on_unions_user_and_team_repos(
-        self, _mock_feature_enabled, mock_user_github_class, mock_team_github_class
-    ):
-        _create_user_github_integration(self.user)
-        self._create_team_github_integration()
-
-        mock_user_github = MagicMock()
-        mock_user_github.list_all_cached_repositories.return_value = [
-            _repo_dict("posthog", "posthog-js", 1),
-            _repo_dict("posthog", "shared", 2),
-        ]
-        mock_user_github_class.return_value = mock_user_github
-
-        mock_team_github = MagicMock()
-        mock_team_github.list_all_cached_repositories.return_value = [
-            _repo_dict("posthog", "posthog", 3),
-            _repo_dict("posthog", "shared", 2),
-        ]
-        mock_team_github_class.return_value = mock_team_github
-
-        result = _get_full_repo_names(self.slack_integration, user_id=self.user.id)
-
-        assert result == ["posthog/posthog", "posthog/posthog-js", "posthog/shared"]
-
-    @patch("products.slack_app.backend.api.posthoganalytics.feature_enabled", return_value=True)
-    def test_flag_on_no_personal_install_returns_team_repos(
-        self, _mock_feature_enabled, mock_user_github_class, mock_team_github_class
-    ):
-        self._create_team_github_integration()
-
-        mock_team_github = MagicMock()
-        mock_team_github.list_all_cached_repositories.return_value = [
-            _repo_dict("posthog", "posthog", 1),
-            _repo_dict("posthog", "plugin-server", 2),
-        ]
-        mock_team_github_class.return_value = mock_team_github
-
-        result = _get_full_repo_names(self.slack_integration, user_id=self.user.id)
-
-        assert result == ["posthog/plugin-server", "posthog/posthog"]
-        mock_user_github_class.assert_not_called()
-
-
 @patch("products.slack_app.backend.api.UserGitHubIntegration")
 class TestGetFullRepoNamesCache:
     @pytest.fixture(autouse=True)
